add some validation to delete endpoint.

+++ routes/messages.js
@@ -1,14 +1,27 @@
 const express = require("express");
 const { RouteConfigs } = require("../configs/routes-configs");
+const ObjectId = require("mongoose").Types.ObjectId;
 
 const MessagesController = (messageRepo) => {
   const router = express.Router();
   const uri = RouteConfigs.getMessagesUri();
 
-  router.delete(uri + "/" + ":id", (req, resp, next) => {
+  router.delete(uri + "/" + ":id", async (req, resp, next) => {
     const id = req.params.id;
 
-    messageRepo.deleteById(id);
+    if (!ObjectId.isValid(id)) {
+      resp.status(400);
+      resp.send("Bad request");
+      return;
+    }
+
+    const result = await messageRepo.deleteById(id);
+
+    if (result.ok !== 1 || result.deleteCount !== 1) {
+      resp.status(404);
+      resp.send("Not found");
+      return;
+    }
 
     next();
   });
